pyproject.toml
[tool.poetry]
name = "astrobot"
version = "22.1.5"
description = "the all-encompassing god-level bot for mochjicord"
authors = ["kevinshome <noah.tanner7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.26"
psycopg2-binary = "^2.9.1"
py-cord = {git = "https://github.com/Pycord-Development/pycord.git"}
python-dateutil = "^2.8.2"
tekore = "^4.1.0"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"

[build-system]
requires = ["poetry>=1.1.11"]
build-backend = "poetry.masonry.api"
